feat(backend/temp_result_conv): リザルト登録を賢くする + ResultLog も登録する

+++ backend/app/models/temporary_result_converter/processor.rb
@@ -14,19 +14,31 @@ module TemporaryResultConverter
         next unless musics.key?(title)
 
         result_sets = results.map do |r|
-          [
-            r.id,
-            to_result(r, musics[title].find_map(r.play_style, r.difficulty)),
-          ]
-        end.to_h.compact
+          map = musics[title].find_map(r.play_style, r.difficulty)
+          next if map.nil?
 
-        ids = result_sets.keys
-        results = result_sets.values
+          [r.id, to_result(r, map), to_result_log(r, map)]
+        end.compact
 
-        results.each { |r| r.run_callbacks(:save) }
-        Result.import(results)
+        next if result_sets.empty?
 
-        TemporaryResult.where(id: ids).delete_all
+        ids, results, result_logs = result_sets.transpose.map(&:compact)
+
+        unless results.empty?
+          results = uniq_results(results)
+
+          results.each { |r| r.run_callbacks(:save) }
+          Result.import(results)
+        end
+
+        unless result_logs.empty?
+          result_logs = uniq_result_logs(result_logs)
+
+          result_logs.each { |r| r.run_callbacks(:save) }
+          ResultLog.import(result_logs)
+        end
+
+        TemporaryResult.where(id: ids).delete_all unless ids.empty?
       end
     end
 
@@ -48,8 +60,26 @@ module TemporaryResultConverter
         end
     end
 
+    def current_results
+      @current_results ||=
+        begin
+          records = Result.where(map: musics.flat_map(&:maps))
+          records.group_by(&:map_id)
+        end
+    end
+
+    def find_result(user_id:, map:)
+      key = [user_id, map.id]
+      @result_cache ||= {}
+
+      return @result_cache[key] if @result_cache.key?(key)
+
+      @result_cache[key] = Result.find_by(user_id: user_id, map: map)
+    end
+
     def to_result(temp_result, map)
-      return nil if map.nil?
+      current_result = find_result(user_id: temp_result.user_id, map: map)
+      return unless current_result.nil?
 
       Result.new(
         user_id: temp_result.user_id,
@@ -60,6 +90,51 @@ module TemporaryResultConverter
         clear_lamp: temp_result.clear_lamp,
         last_played_at: temp_result.last_played_at,
       )
+    end
+
+    def find_result_log(user_id:, map:, last_played_at:)
+      key = [user_id, map.id, last_played_at]
+      @result_log_cache ||= {}
+
+      return @result_log_cache[key] if @result_log_cache.key?(key)
+
+      @result_log_cache[key] =
+        ResultLog.find_by(user_id: user_id, map: map, last_played_at: last_played_at)
+    end
+
+    def to_result_log(temp_result, map)
+      current_result_log = find_result_log(user_id: temp_result.user_id, map: map, last_played_at: temp_result.last_played_at)
+      return if !current_result_log.nil? && current_result_log.last_played_at <= temp_result.last_played_at
+
+      result_log = ResultLog.new(
+        user_id: temp_result.user_id,
+        result_batch_id: temp_result.result_batch_id,
+        map: map,
+        score: temp_result.score,
+        miss_count: temp_result.miss_count,
+        clear_lamp: temp_result.clear_lamp,
+        last_played_at: temp_result.last_played_at,
+      )
+
+      result_log
+    end
+
+    def uniq_results(results)
+      results.sort_by(&:last_played_at).reverse.uniq { |r| [r.user_id, r.map_id] }
+    end
+
+    def uniq_result_logs(result_logs)
+      [].tap do |arr|
+        result_logs.group_by { |r| [r.user_id, r.map_id] }.each_value do |rs|
+          ([nil] + rs.sort_by(&:last_played_at)).each_cons(2) do |old, new|
+            if old.nil?
+              arr << new
+            elsif old < new
+              arr << new
+            end
+          end
+        end
+      end
     end
   end
 end
